requirements.txt
flask>=2.3.3
requests>=2.31.0
python-dotenv>=1.0.0
coverage>=7.3.2
pytest>=7.4.3
pytest-cov>=4.1.0
gitpython>=3.1.40
pygithub>=2.1.1
jinja2>=3.1.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
markdown>=3.5.1
pyyaml>=6.0.1
click>=8.1.7
gunicorn>=21.2.0
aiohttp>=3.9.0
asyncio>=3.4.3
openai>=1.0.0